refactor net classes
